requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.2
python-magic==0.4.27
aiofiles==23.2.1
jinja2==3.1.2
python-slugify==8.0.1
emails==0.6
dnspython==2.4.2